+++ src/htmx/index.js
@@ -2,8 +2,8 @@ const HtmlRenderer = require('./HtmlRenderer');
 const LiveHtmlServer = require('./LiveHtmlServer');
 const LivePage = require('./LivePage');
 const { attribute, codeBlock, each, html, safeUrl: url } = require('./Html');
-const { action, page, state, view } = require('./metadata');
+const { action, component, page, state, view } = require('./metadata');
 const { start } = require('./start');
 const { exportStatic } = require('./StaticExporter');
 
-module.exports = { action, attribute, codeBlock, each, exportStatic, html, HtmlRenderer, LiveHtmlServer, LivePage, page, start, state, url, view };
+module.exports = { action, attribute, codeBlock, component, each, exportStatic, html, HtmlRenderer, LiveHtmlServer, LivePage, page, start, state, url, view };